requirements.txt
discord.py-self
asyncio